refactor(provider): consolidate feature wiring configs
