--- backend/src/main/java/com/alligator/market/backend/provider/adapter/moex/iss/config/MoexIssProviderConfig.java
@@ -1,35 +1,0 @@
-package com.alligator.market.backend.provider.adapter.moex.iss.config;
-
-import com.alligator.market.backend.provider.adapter.moex.iss.MoexIssProvider;
-import com.alligator.market.backend.provider.adapter.moex.iss.config.handlers.MoexIssHandlersConfig;
-import com.alligator.market.domain.instrument.Instrument;
-import com.alligator.market.domain.provider.handler.InstrumentHandler;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-import java.util.Set;
-
-/**
- * Wiring-конфигурация {@link MoexIssProvider}.
- */
-@Configuration(proxyBeanMethods = false)
-@Import(MoexIssHandlersConfig.class)
-public class MoexIssProviderConfig {
-
-    public static final String BEAN_NAME = "moexIssProvider";
-
-    /**
-     * Бин {@link MoexIssProvider}.
-     *
-     * @param handlers набор обработчиков инструментов провайдера MOEX ISS
-     */
-    @Bean(BEAN_NAME)
-    public MoexIssProvider moexIssProvider(
-            @Qualifier(MoexIssHandlersConfig.BEAN_NAME)
-            Set<InstrumentHandler<MoexIssProvider, ? extends Instrument>> handlers
-    ) {
-        return new MoexIssProvider(handlers);
-    }
-}
