+++ backend/src/main/java/com/alligator/market/backend/provider/config/adapter/moex/iss/handlers/MoexIssHandlersConfig.java
@@ -1,0 +1,38 @@
+package com.alligator.market.backend.provider.config.adapter.moex.iss.handlers;
+
+import com.alligator.market.backend.provider.adapter.moex.iss.MoexIssProvider;
+import com.alligator.market.backend.provider.config.adapter.moex.iss.instrument.forex.spot.handler.MoexIssFxSpotHandlerConfig;
+import com.alligator.market.backend.provider.adapter.moex.iss.instrument.forex.spot.handler.MoexIssFxSpotHandler;
+import com.alligator.market.domain.instrument.Instrument;
+import com.alligator.market.domain.provider.handler.InstrumentHandler;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+import java.util.Set;
+
+/**
+ * Wiring-конфигурация набора всех обработчиков провайдера MOEX ISS.
+ */
+@Configuration(proxyBeanMethods = false)
+@Import({
+        MoexIssFxSpotHandlerConfig.class
+        // + другие *HandlerConfig по мере добавления
+})
+public class MoexIssHandlersConfig {
+
+    public static final String BEAN_NAME = "moexIssHandlers";
+
+    /**
+     * Бин набора всех обработчиков MOEX ISS.
+     *
+     * @param fxSpotHandler обработчик инструмента FOREX_SPOT
+     */
+    @Bean(BEAN_NAME)
+    public Set<InstrumentHandler<MoexIssProvider, ? extends Instrument>> moexIssHandlers(
+            MoexIssFxSpotHandler fxSpotHandler
+            // + другие *Handler по мере добавления
+    ) {
+        return Set.of(fxSpotHandler);
+    }
+}
